+++ CHANGELOG.md
@@ -84,6 +84,7 @@
 * [BUGFIX]: Ingester: Panic when push and read reactive limiters are enabled with prioritization. #13482
 * [BUGFIX] Query-frontend: Fix excessive CPU and memory consumption when running sharding inside MQE. #13580
 * [BUGFIX] Query-frontend: Fix incorrect query results when running sharding inside MQE is enabled and the query contains a subquery eligible for subquery spin-off wrapped in a shardable aggregation. #13619
+* [BUGFIX] Query-frontend: Fix silent panic when executing a remote read API request if the request has no matchers. #13745
 
 ### Mixin